+++ service/src/main/java/com/nosota/mwallet/service/SystemStatisticService.java
@@ -3,6 +3,7 @@ package com.nosota.mwallet.service;
 import com.nosota.mwallet.repository.SystemStatisticRepository;
 import jakarta.transaction.Transactional;
 import lombok.AllArgsConstructor;
+import lombok.extern.slf4j.Slf4j;
 import org.springframework.stereotype.Service;
 import org.springframework.validation.annotation.Validated;
 
@@ -11,6 +12,7 @@ import java.math.BigDecimal;
 @Service
 @Validated
 @AllArgsConstructor
+@Slf4j
 public class SystemStatisticService
 {
     private final SystemStatisticRepository systemStatisticRepository;