+++ service/src/main/java/com/nosota/mwallet/service/TransactionHistoryService.java
@@ -8,6 +8,7 @@ import jakarta.persistence.*;
 import jakarta.validation.constraints.NotNull;
 import jakarta.validation.constraints.Positive;
 import lombok.AllArgsConstructor;
+import lombok.extern.slf4j.Slf4j;
 import org.springframework.stereotype.Service;
 import org.springframework.validation.annotation.Validated;
 
@@ -20,6 +21,7 @@ import java.util.stream.Collectors;
 @Service
 @Validated
 @AllArgsConstructor
+@Slf4j
 public class TransactionHistoryService {
 
     @PersistenceContext