+++ service/src/main/java/com/nosota/mwallet/service/TransactionService.java
@@ -16,6 +16,7 @@ import jakarta.validation.constraints.NotEmpty;
 import jakarta.validation.constraints.NotNull;
 import jakarta.validation.constraints.Positive;
 import lombok.AllArgsConstructor;
+import lombok.extern.slf4j.Slf4j;
 import org.springframework.dao.DataAccessException;
 import org.springframework.stereotype.Service;
 import org.springframework.validation.annotation.Validated;
@@ -26,6 +27,7 @@ import java.util.UUID;
 @Service
 @Validated
 @AllArgsConstructor
+@Slf4j
 public class TransactionService {
 
     private final WalletService walletService;