+++ service/src/main/java/com/nosota/mwallet/service/WalletBalanceService.java
@@ -6,6 +6,7 @@ import jakarta.persistence.Query;
 import jakarta.transaction.Transactional;
 import jakarta.validation.constraints.NotNull;
 import lombok.AllArgsConstructor;
+import lombok.extern.slf4j.Slf4j;
 import org.springframework.stereotype.Service;
 import org.springframework.validation.annotation.Validated;
 
@@ -14,6 +15,7 @@ import java.math.BigDecimal;
 @Service
 @Validated
 @AllArgsConstructor
+@Slf4j
 public class WalletBalanceService {
     private final EntityManager entityManager;